[apiConformance] Split core_integration
