+++ src/plugins/template/tests/functional/shared_tests_instances/behavior/ov_plugin/properties_tests.cpp
@@ -11,21 +11,8 @@ namespace {
 
 const std::vector<ov::AnyMap> inproperties = {
         {ov::device::id("UNSUPPORTED_DEVICE_ID_STRING")},
+        {{"unsupported_key", "4"}}
 };
-
-const std::vector<ov::AnyMap> hetero_inproperties = {
-        {ov::device::id("UNSUPPORTED_DEVICE_ID_STRING")},
-};
-
-const std::vector<ov::AnyMap> multi_inproperties = {
-        {ov::device::id("UNSUPPORTED_DEVICE_ID_STRING")},
-};
-
-
-const std::vector<ov::AnyMap> auto_inproperties = {
-        {ov::device::id("UNSUPPORTED_DEVICE_ID_STRING")},
-};
-
 
 const std::vector<ov::AnyMap> auto_batch_inproperties = {
         {ov::device::id("UNSUPPORTED_DEVICE_ID_STRING")},
@@ -33,26 +20,9 @@ const std::vector<ov::AnyMap> auto_batch_inproperties = {
 
 INSTANTIATE_TEST_SUITE_P(DISABLED_smoke_BehaviorTests, OVPropertiesIncorrectTests,
                         ::testing::Combine(
-                                ::testing::Values(CommonTestUtils::DEVICE_TEMPLATE),
+                                ::testing::Values(CommonTestUtils::DEVICE_TEMPLATE, CommonTestUtils::DEVICE_HETERO, 
+                                                  CommonTestUtils::DEVICE_MULTI, CommonTestUtils::DEVICE_AUTO),
                                 ::testing::ValuesIn(inproperties)),
-                        OVPropertiesIncorrectTests::getTestCaseName);
-
-INSTANTIATE_TEST_SUITE_P(DISABLED_smoke_Hetero_BehaviorTests, OVPropertiesIncorrectTests,
-                        ::testing::Combine(
-                                ::testing::Values(CommonTestUtils::DEVICE_HETERO),
-                                ::testing::ValuesIn(hetero_inproperties)),
-                        OVPropertiesIncorrectTests::getTestCaseName);
-
-INSTANTIATE_TEST_SUITE_P(DISABLED_smoke_Multi_BehaviorTests, OVPropertiesIncorrectTests,
-                        ::testing::Combine(
-                                ::testing::Values(CommonTestUtils::DEVICE_MULTI),
-                                ::testing::ValuesIn(multi_inproperties)),
-                        OVPropertiesIncorrectTests::getTestCaseName);
-
-INSTANTIATE_TEST_SUITE_P(DISABLED_smoke_Auto_BehaviorTests, OVPropertiesIncorrectTests,
-                        ::testing::Combine(
-                                ::testing::Values(CommonTestUtils::DEVICE_AUTO),
-                                ::testing::ValuesIn(auto_inproperties)),
                         OVPropertiesIncorrectTests::getTestCaseName);
 
 INSTANTIATE_TEST_SUITE_P(DISABLED_smoke_AutoBatch_BehaviorTests, OVPropertiesIncorrectTests,
@@ -117,4 +87,41 @@ INSTANTIATE_TEST_SUITE_P(DISABLED_smoke_AutoBatch_BehaviorTests, OVPropertiesTes
                 ::testing::Values(CommonTestUtils::DEVICE_BATCH),
                 ::testing::ValuesIn(auto_batch_properties)),
         OVPropertiesTests::getTestCaseName);
+
+
+//
+// OV Class GetMetric
+//
+
+const std::vector<ov::AnyMap> ro_property_template_plugin = {
+        {{ov::PropertyName(ov::supported_properties.name(), ov::supported_properties.mutability), nullptr}},
+        {{ov::PropertyName(ov::available_devices.name(), ov::available_devices.mutability), nullptr}},
+        {{ov::PropertyName(ov::device::full_name.name(), ov::device::full_name.mutability), nullptr}},
+        {{ov::PropertyName(ov::device::capabilities.name(), ov::device::capabilities.mutability), nullptr}}
+};
+
+INSTANTIATE_TEST_SUITE_P(
+        smoke_OVCheckChangePropComplieModleGetPropTestsRO, OVCheckChangePropComplieModleGetPropTestsRO,
+        ::testing::Combine(::testing::Values(CommonTestUtils::DEVICE_TEMPLATE),
+                           ::testing::ValuesIn(ro_property_template_plugin)));
+
+INSTANTIATE_TEST_SUITE_P(
+        smoke_OVClassGetMetricTest, OVClassGetMetricTest_RANGE_FOR_ASYNC_INFER_REQUESTS,
+        ::testing::Values(CommonTestUtils::DEVICE_TEMPLATE));
+
+// 
+// Check set of incorrect properties 
+//
+std::vector<ov::AnyMap> incorrect_properies = {{{"unsupported_key", "4"}}};
+
+INSTANTIATE_TEST_SUITE_P(
+        smoke_OVGetPropertiesIncorrectTests, OVGetPropertiesIncorrectTests,
+        ::testing::Combine(::testing::Values(CommonTestUtils::DEVICE_TEMPLATE),
+                           ::testing::ValuesIn(incorrect_properies)));
+
+INSTANTIATE_TEST_SUITE_P(
+        smoke_OVClassGetAvailableDevices, OVClassGetAvailableDevices,
+        ::testing::Values(CommonTestUtils::DEVICE_TEMPLATE));
+
+
 } // namespace
